integration tests: move VLAN integration tests to a separate file
